+++ mithril-test-lab/mithril-end-to-end/src/end_to_end_spec.rs
@@ -1,14 +1,40 @@
 use crate::assertions;
 use crate::MithrilInfrastructure;
-use mithril_common::StdResult;
+use mithril_common::{
+    entities::{Epoch, SignedEntityTypeDiscriminants},
+    StdResult,
+};
 
 pub struct Spec<'a> {
     pub infrastructure: &'a mut MithrilInfrastructure,
+    is_signing_cardano_transactions: bool,
+    is_signing_cardano_stake_distribution: bool,
+    next_era: Option<String>,
+    regenesis_on_era_switch: bool,
 }
 
 impl<'a> Spec<'a> {
-    pub fn new(infrastructure: &'a mut MithrilInfrastructure) -> Self {
-        Self { infrastructure }
+    pub fn new(
+        infrastructure: &'a mut MithrilInfrastructure,
+        signed_entity_types: Vec<String>,
+        next_era: Option<String>,
+        regenesis_on_era_switch: bool,
+    ) -> Self {
+        Self {
+            infrastructure,
+            is_signing_cardano_transactions: signed_entity_types.contains(
+                &SignedEntityTypeDiscriminants::CardanoTransactions
+                    .as_ref()
+                    .to_string(),
+            ),
+            is_signing_cardano_stake_distribution: signed_entity_types.contains(
+                &SignedEntityTypeDiscriminants::CardanoStakeDistribution
+                    .as_ref()
+                    .to_string(),
+            ),
+            next_era,
+            regenesis_on_era_switch,
+        }
     }
 
     pub async fn run(&mut self) -> StdResult<()> {
@@ -71,6 +97,45 @@ impl<'a> Spec<'a> {
         )
         .await?;
 
+        // Verify that artifacts are produced and signed correctly
+        let mut target_epoch = self.verify_artifacts_production(target_epoch).await?;
+
+        // Verify that artifacts are produced and signed correctly after era switch
+        if let Some(next_era) = &self.next_era {
+            // Switch to next era
+            self.infrastructure
+                .register_switch_to_next_era(next_era)
+                .await?;
+            target_epoch += 5;
+            assertions::wait_for_target_epoch(
+                self.infrastructure.chain_observer(),
+                target_epoch,
+                "epoch after which the era switch will have triggered".to_string(),
+            )
+            .await?;
+
+            // Proceed to a re-genesis of the certificate chain
+            if self.regenesis_on_era_switch {
+                assertions::bootstrap_genesis_certificate(self.infrastructure.aggregator_mut())
+                    .await?;
+                target_epoch += 5;
+                assertions::wait_for_target_epoch(
+                    self.infrastructure.chain_observer(),
+                    target_epoch,
+                    "epoch after which the re-genesis on era switch will be completed".to_string(),
+                )
+                .await?;
+            }
+
+            // Verify that artifacts are produced and signed correctly
+            self.verify_artifacts_production(target_epoch).await?;
+        }
+
+        Ok(())
+    }
+
+    async fn verify_artifacts_production(&self, target_epoch: Epoch) -> StdResult<Epoch> {
+        let aggregator_endpoint = self.infrastructure.aggregator().endpoint();
         let expected_epoch_min = target_epoch - 3;
         // Verify that mithril stake distribution artifacts are produced and signed correctly
         {
@@ -116,7 +181,7 @@ impl<'a> Spec<'a> {
         }
 
         // Verify that Cardano transactions artifacts are produced and signed correctly
-        if self.infrastructure.is_signing_cardano_transactions() {
+        if self.is_signing_cardano_transactions {
             let hash = assertions::assert_node_producing_cardano_transactions(&aggregator_endpoint)
                 .await?;
             let certificate_hash = assertions::assert_signer_is_signing_cardano_transactions(
@@ -143,7 +208,7 @@ impl<'a> Spec<'a> {
         }
 
         // Verify that Cardano stake distribution artifacts are produced and signed correctly
-        if self.infrastructure.is_signing_cardano_stake_distribution() {
+        if self.is_signing_cardano_stake_distribution {
             {
                 let (hash, epoch) = assertions::assert_node_producing_cardano_stake_distribution(
                     &aggregator_endpoint,
@@ -173,6 +238,6 @@ impl<'a> Spec<'a> {
             }
         }
 
-        Ok(())
+        Ok(target_epoch)
     }
 }