+++ mithril-test-lab/mithril-end-to-end/src/main.rs
@@ -72,6 +72,14 @@ pub struct Args {
     #[clap(long, default_value = "thales")]
     mithril_era: String,
 
+    /// Mithril next era to run
+    #[clap(long)]
+    mithril_next_era: Option<String>,
+
+    /// Mithril re-genesis on era switch (used only when 'mithril_next_era' is set)
+    #[clap(long, default_value_t = false)]
+    mithril_era_regenesis_on_switch: bool,
+
     /// Mithril era reader adapter
     #[clap(long, default_value = "cardano-chain")]
     mithril_era_reader_adapter: String,
@@ -191,10 +199,11 @@ async fn main() -> StdResult<()> {
         mithril_run_interval: args.mithril_run_interval,
         mithril_era: args.mithril_era,
         mithril_era_reader_adapter: args.mithril_era_reader_adapter,
-        signed_entity_types: args.signed_entity_types,
+        signed_entity_types: args.signed_entity_types.clone(),
         run_only_mode,
         use_p2p_network_mode,
         use_p2p_passive_relays,
+        use_era_specific_work_dir: args.mithril_next_era.is_some(),
     })
     .await?;
 
@@ -204,7 +213,12 @@ async fn main() -> StdResult<()> {
             run_only.start().await
         }
         false => {
-            let mut spec = Spec::new(&mut infrastructure);
+            let mut spec = Spec::new(
+                &mut infrastructure,
+                args.signed_entity_types,
+                args.mithril_next_era,
+                args.mithril_era_regenesis_on_switch,
+            );
             spec.run().await
         }
     };