+++ mithril-test-lab/mithril-end-to-end/src/mithril/infrastructure.rs
@@ -3,14 +3,13 @@ use crate::{
     RelayPassive, RelaySigner, Signer, DEVNET_MAGIC_ID,
 };
 use mithril_common::chain_observer::{ChainObserver, PallasChainObserver};
-use mithril_common::entities::{PartyId, ProtocolParameters, SignedEntityTypeDiscriminants};
+use mithril_common::entities::{Epoch, PartyId, ProtocolParameters};
 use mithril_common::{CardanoNetwork, StdResult};
 use slog_scope::info;
 use std::borrow::BorrowMut;
+use std::fs;
 use std::path::PathBuf;
 use std::sync::Arc;
-use std::time::Duration;
-use tokio::time::sleep;
 
 use super::signer::SignerConfig;
 
@@ -27,6 +26,7 @@ pub struct MithrilInfrastructureConfig {
     pub run_only_mode: bool,
     pub use_p2p_network_mode: bool,
     pub use_p2p_passive_relays: bool,
+    pub use_era_specific_work_dir: bool,
 }
 
 pub struct MithrilInfrastructure {
@@ -40,8 +40,9 @@ pub struct MithrilInfrastructure {
     relay_passives: Vec<RelayPassive>,
     cardano_chain_observer: Arc<dyn ChainObserver>,
     run_only_mode: bool,
-    is_signing_cardano_transactions: bool,
-    is_signing_cardano_stake_distribution: bool,
+    current_era: String,
+    era_reader_adapter: String,
+    use_era_specific_work_dir: bool,
 }
 
 impl MithrilInfrastructure {
@@ -86,17 +87,49 @@ impl MithrilInfrastructure {
             relay_passives,
             cardano_chain_observer,
             run_only_mode: config.run_only_mode,
-            is_signing_cardano_transactions: config.signed_entity_types.contains(
-                &SignedEntityTypeDiscriminants::CardanoTransactions
-                    .as_ref()
-                    .to_string(),
-            ),
-            is_signing_cardano_stake_distribution: config.signed_entity_types.contains(
-                &SignedEntityTypeDiscriminants::CardanoStakeDistribution
-                    .as_ref()
-                    .to_string(),
-            ),
+            current_era: config.mithril_era.clone(),
+            era_reader_adapter: config.mithril_era_reader_adapter.clone(),
+            use_era_specific_work_dir: config.use_era_specific_work_dir,
         })
+    }
+
+    async fn register_startup_era(
+        aggregator: &mut Aggregator,
+        config: &MithrilInfrastructureConfig,
+    ) -> StdResult<()> {
+        let era_epoch = Epoch(0);
+        if config.mithril_era_reader_adapter == "cardano-chain" {
+            assertions::register_era_marker(
+                aggregator,
+                &config.devnet,
+                &config.mithril_era,
+                era_epoch,
+            )
+            .await?;
+        }
+
+        Ok(())
+    }
+
+    pub async fn register_switch_to_next_era(&mut self, next_era: &str) -> StdResult<()> {
+        let next_era_epoch = self
+            .chain_observer()
+            .get_current_epoch()
+            .await?
+            .unwrap_or_default()
+            + 1;
+        if self.era_reader_adapter == "cardano-chain" {
+            assertions::register_era_marker(
+                &mut self.aggregator,
+                &self.devnet,
+                next_era,
+                next_era_epoch,
+            )
+            .await?;
+        }
+        self.current_era = next_era.to_owned();
+
+        Ok(())
     }
 
     async fn start_aggregator(
@@ -124,11 +157,9 @@ impl MithrilInfrastructure {
             m: 105,
             phi_f: 0.95,
         });
-        if config.mithril_era_reader_adapter == "cardano-chain" {
-            assertions::register_era_marker(&mut aggregator, &config.devnet, &config.mithril_era)
-                .await?;
-            sleep(Duration::from_secs(5)).await;
-        }
+
+        Self::register_startup_era(&mut aggregator, config).await?;
+
         aggregator.serve()?;
 
         Ok(aggregator)
@@ -224,6 +255,7 @@ impl MithrilInfrastructure {
             };
 
             let mut signer = Signer::new(&SignerConfig {
+                signer_number: index + 1,
                 aggregator_endpoint,
                 pool_node,
                 cardano_cli_path: &config.devnet.cardano_cli_path(),
@@ -280,19 +312,22 @@ impl MithrilInfrastructure {
     }
 
     pub fn build_client(&self) -> StdResult<Client> {
-        Client::new(self.aggregator.endpoint(), &self.work_dir, &self.bin_dir)
+        let work_dir = if self.use_era_specific_work_dir {
+            let era_work_dir = self.work_dir.join(format!("era.{}", self.current_era));
+            if !era_work_dir.exists() {
+                fs::create_dir(&era_work_dir)?;
+            }
+
+            era_work_dir
+        } else {
+            self.work_dir.clone()
+        };
+
+        Client::new(self.aggregator.endpoint(), &work_dir, &self.bin_dir)
     }
 
     pub fn run_only_mode(&self) -> bool {
         self.run_only_mode
-    }
-
-    pub fn is_signing_cardano_transactions(&self) -> bool {
-        self.is_signing_cardano_transactions
-    }
-
-    pub fn is_signing_cardano_stake_distribution(&self) -> bool {
-        self.is_signing_cardano_stake_distribution
     }
 
     pub async fn tail_logs(&self, number_of_line: u64) -> StdResult<()> {